Resolves #2203: Allow for setting the store state cacheability during checkVersion (#2204)
A new record store configuration parameter is added. It can be configured to instruct `checkVersion` to update the `cacheability` flag. If the store header is alreadyin the desired state, then no update is performed. In addition, this allows the user to skip updating the flag if the store is pre-existing, which was the old behavior, and which is still important during roll outs. In particular, if the same set of instances are configured to operate on the same store, and some of them are configured to set cacheability to `true` while others to set it to `false`, then operations from the one set will always step on operations from the other set. To avoid that, one can first set the parameter to ignore existing stores and then make a second change to start updating existing stores again.

This resolve #2203.