([Wave 1] Wrong number of recourd counts returned by recordIdOrUid query) (REL-499)
This adds few tests for SQL statements that contains disjunction predicates which might be turned into union and verifies that the result set must not contain any (extra) rows. It also includes an improvement regarding the output of the YAML plan (mis)match result; it is rendered now as a colored diff making it easier to quickly detect plan difference instead of manually matching plan strings.