Preserve .cfs file reference when deleting
It's possible that will lucene will delete the .cfs, but not
.si, or .cfe, but we put the contents/reference for .si/.cfe
on the file reference for .cfs. This means that we will have
a corrupted index when the .cfs is deleted.
As a quick fix I changed the file reference for .cfs to store
whether or not it had been deleted, so it will stick around
preserving the .si and .cfe. This isn't great though because
it will stick around forever, and it's confusing.