Don't wrap IOException in RecordCoreArgumentException
Most importantly, Lucene's logic has a bunch of handling for
IOException, but by wrapping it in a non-IOException, the logic
doesn't work, and the exception bubbles all the way up.
This does mean that we won't add the SOURCE_FILE logging key,
but that's probably better than breaking lucene.

We may want to create our own IOException that can add log info,
but not write now...