Completely remove the use of special schema keyspace
It's possible that this would never manifest in real record-layer
use cases, but IndexWriter has methods to add another index from
another directory. It does so by copying the segments as-is.
Within a directory we maintain a global mapping between fieldname
and fieldnumber, but independent directories might not do that, and
thus, we could end up with different segments having different
mappings. This breaks the performance optimization of storing *.fnm
in a separate range keyed by the bitset of field numbers, since
the same bitset could point to a different set of mappings.

This change puts the *.fnm content directly on the File Reference,
which is more expensive storage-wise, and on the initial listAllFiles
rangeRead, because we duplicate all the data in these identical,
or nearly identical files.

It is possible that we could find an alternate approach that still
respects Lucene's edge cases by detecting duplicate .fnm files and
storing them in a central place. Normal record-layer uses would
generally just reference that shared location, but calling
addIndexes would result in more .fnm files being serialized.

Note: the rather low-level place at which it copies segments as-is
might make it hard to detect, i.e. we would have to provide an
implementation of IndexInput for .fnm that can find the shared
data and utilize it.