Fixes #2275: Version mutations aren't cleared out by delete records where
The main record store and index maintainers now clear out uncommitted version mutations when performing a record range delete. Previously, this failure to do so meant that it would be dangerous to perform `deleteRecordsWhere` with record versions enabled if there were uncommitted record saves performed in the same transaction. Committing the transaction would lead to the record data not being written, but the version information and the version indexes would be, leading to corrupt record stores and orphaned index entries. This cleans up that data during the clear so that we don't have uncommitted versions.

There's something to be said about the abstractions here potentially being a little off, and that we should improve some of our I/O methods here to make this bug harder to write. For now, I've just tried to patch the bug.

This fixes #2275.