Lucene: Store simple files as content on File Reference
Resolves #2295
Resolves #2284

This resolves two issues by storing the .fnm, .cfe and .si on the
LuceneFileReference as `content`.
Previously it would store .si and .cfe directly on the reference for
.cfs.
It would also try to deduplicate the .fnm information, and store that
in a central place, now it is stored as content on the .fnm file.

As a consequence, this removes LuceneOptimizedWrappedDirectory,
LuceneOptimizedSegmentInfoFormat and LuceneOptimizedFieldInfosFormat
as they were focused on special casing this behavior.