Fix performance issues with previous commit
The change to store `.fnm` on the file reference caused a
noticeable performance regression.
What was happening is that when it went to create the `.cfs`,
it would copy the `.fnm` content out of the file reference
and into the `.cfs`. So anytime it had to lookup the `.fnm`
it would have to do an additional bock read into the `.cfs`,
which effecively doubled the number of reads done during
inserts.

This change moves the content out of the file reference and
into a separate keyspace. It then, does some reference
counting so that when the `.cfs` is created it adds a
"pointer" to the content from the associated `.cfe`.

This tries to workaround the performance cost of loading all the
field infos, by doing it as a single range scan.
The approach of using ids, and reference counting, I think means
that we could de-duplicate the fieldInfos, potentially using the
bitset from before combined, so that as long as a segment is a
subset of some FieldInfos, it can reuse it.