Introduce Relational structured query API (REL-564)
This introduces Relational Structured Query API, it enables constructing a SQL query in statically-typed fashion either from scratch or from a given SQL statement that is syntactically (and to some degree, semantically) valid. The later approach enables traversing a SQL statement and examining it in a structured way, this enables for example checking accessed fields of a select statement to apply ad-hoc security principles.
It also enables decorating the query and effectively add extra layers of business logic to customer queries such as appending extra predicates to the where clause of a select statement e.g. to avoid returning expired records, or setting extra fields in an update statement for accounting and quota management.