Store a global FieldInfos
This adds a single global field infos, that can be reused for all
segments, as long as they have a subset of the same name-to-number
mapping.
This is similar to what there currently is in main, but if there
are segments with incompatible name-to-number mappings, it handles
that. Hopefully this change will allow us to preserve the performance,
and be correct.