Resolves #2336: PlanOrderingKey.mergedComparisonKey too strict about fields between requested ordering and primary key
This adjusts the way the plan ordering key is constructed for unions and intersections so that it can accept additional columns in the order. Effectively, in the case where you have an index on `[a, b, c, d]` with an equality on `a`, a disjunct of equalities on `b`, and a primary key `k` (the hypothetical raised in the original issue), then this will construct an ordered union between multiple scans of the index (one leg for each member of the disjunction on `b`) and will use the comparison key `[c, d, k]` to ensure that the results are processed in an expected order.

This resolves #2336.