Fixes #2353: DualPlannerTests are only running old planner if also marked as parameterized
This updates the `DualPlannerTest` extension so that parameterized tests will run with both planners. This adjusts the work done in #2022 so that the Cascades planner is configured during test execution, which is what was missing with the previous approach. The rest of the changes here are in response to that, mainly fixing up the `FDBOrQueryToUnionTest` so that (1) tests have nice names that are easier to debug and (2) plan assertions that were failing are fixed.

This fixes #2353.