Some cleanup from draft PR
I opened PR #2296, and is commit addressed issues found there
1. Some cleanup just by my own self-review
2. Cleanup in response to automatic checks
3. Fix assertionsfor LuceneOnlineIndexingTest. In my previous
rebases, I did not notice that this change broke this test.
With a previous expected file count of 12, I would expect it
to be 34 now. Some of the tests didn't fail though, so I
fixed them so that they all generate some segments, and gave
each test a unique max number of segments.