Fix assertions for LuceneOnlineIndexingTest
The change in this PR increased the count of files here, because
we now have files for .cfe and .si. I noticed that only some of
the tests were failing, which was suspicious, and discovered that
COMPLEX_DOC tests and grouped tests weren't working correctly.
These tests should now be appropriately asserting.