Support delayed nested star expansion (REL-594)
This introduces some changes in the semantics of SQL (*) expansion. It effectively avoids deconstructing the underlying record into individual columns (what we do for *), and instead returns a record holding all the columns.

When this syntax is used, the same query plan can be reused to return results from tables with a different number of columns. This helps us avoiding breaking customer continuations because of a plan-hash change related to an evolving record header for example.

Moreover, customer' `select *` ... queries are adapted to:

Leverage the new `(*)` feature to make the physical plans more stable.
Return the `__ROW_VERSION` as well, as customer expects this field to be returned in this case and set correctly in the resulting record(s).
A new test is added that simulates record-header evolution, it asserts that the plan hash does not change even if the record header changes leveraging the newly added planHash YAML query configuration command.

This fixes ( ([Wave1/2] Adapt plan generator star expansion))