Self-correcting mode for `explain` configurations in YAML (REL-599)
This extends YAML test framework with a new feature that enables it to replace the expected plain explanation with the actual it got from the execution in the test suite file, effectively enabling it to self-adapt the file when necessary.

This could save us hours of tedious work of manually adapting the explain (we have more than a hundred of them in some files) when a systematic change in the plans arise due to an internal change in the optimiser or the plan generator.