Lucene: Store simple files as content on File Reference (#2364)
* Lucene: Store simple files as content on File Reference

Resolves #2295
Resolves #2284

This resolves two issues by storing the .fnm, .cfe and .si on the
LuceneFileReference as `content`.
Previously it would store .si and .cfe directly on the reference for
.cfs.
It would also try to deduplicate the .fnm information, and store that
in a central place, now it is stored as content on the .fnm file.

As a consequence, this removes LuceneOptimizedWrappedDirectory,
LuceneOptimizedSegmentInfoFormat and LuceneOptimizedFieldInfosFormat
as they were focused on special casing this behavior.

* Fix performance issues with previous commit

The change to store `.fnm` on the file reference caused a
noticeable performance regression.
What was happening is that when it went to create the `.cfs`,
it would copy the `.fnm` content out of the file reference
and into the `.cfs`. So anytime it had to lookup the `.fnm`
it would have to do an additional bock read into the `.cfs`,
which effecively doubled the number of reads done during
inserts.

This change moves the content out of the file reference and
into a separate keyspace. It then, does some reference
counting so that when the `.cfs` is created it adds a
"pointer" to the content from the associated `.cfe`.

This tries to workaround the performance cost of loading all the
field infos, by doing it as a single range scan.
The approach of using ids, and reference counting, I think means
that we could de-duplicate the fieldInfos, potentially using the
bitset from before combined, so that as long as a segment is a
subset of some FieldInfos, it can reuse it.

* Store a global FieldInfos

This adds a single global field infos, that can be reused for all
segments, as long as they have a subset of the same name-to-number
mapping.
This is similar to what there currently is in main, but if there
are segments with incompatible name-to-number mappings, it handles
that. Hopefully this change will allow us to preserve the performance,
and be correct.

* Some cleanup from draft PR

I opened PR #2296, and is commit addressed issues found there
1. Some cleanup just by my own self-review
2. Cleanup in response to automatic checks
3. Fix assertionsfor LuceneOnlineIndexingTest. In my previous
rebases, I did not notice that this change broke this test.
With a previous expected file count of 12, I would expect it
to be 34 now. Some of the tests didn't fail though, so I
fixed them so that they all generate some segments, and gave
each test a unique max number of segments.

* Address comments from PR #2364

- Pass in CompoundFormat to LuceneOptimiedCompoundFormat
- Don't filter .cfs out of the si files when creating the compound file
  It looks like this is unnecessary, and just a side effect of when
  we used to put the si/cfe content on the cfs reference
- Use stream to get first element for clarity
- Add some clarifying comments about file names for FieldInfos
- We never open FieldInfoFile on compoundReader, so error
- Comments clarifying global FieldInfos algorithm
- Only update the Global if we are resuing it
- Rework async code in setFieldInfoId for correct timer tracking
- Improve trace log message
- Clarifying comment about updating the FieldInfos cache
- Track creating all types of outputs with WAIT_LUCENE_CREATE_OUTPUT
- Add some general assertions about FieldInfos metadata
- Add new tests of FieldInfos integrity
- Rework helpers in LuceneIndexTest to simplify and make clearer

The new test for deleting FieldInfos has found a bug, to be fixed in
the next commit

* Fix FDBDirectory's deleting of FieldInfos

It was using the file reference's id, not the FieldInfos id to clear
from the FieldInfos subspace, it didn't actually delete anything.

* Have FDBDirectory.getAllFieldInfos call asyncToSync

Before, every usage was calling asyncToSync around it, so this
simplifies the code.