Fixes #2377: NPE during maintenance of synthetic indexes if some types have only disabled indexes (#2379)
* Fixes #2377: NPE during maintenance of synthetic indexes if some types have only disabled indexes

Now, when we go to index a record of a synthetic type, we only go through the process of generating a plan if there is at least one non-disabled index defined on it. Previously, we would generate the plan if there was any index defined on it, and then we would filter out any types without any non-disabled indexes later. The logic for that later filtering could result in a null pointer exception, though, so this updated approach aligns what gets planned with what is actually necessary to execute. This also can save on planning time if there are disabled indexes.

This fixes #2377.