RESOLVES Add Lucene Partitioning Metadata #2390
WIP of the Graceful Degradation Feature

This initial update adds support for partitioning a logical Lucene index into multiple physical indexes spanning a time range.
Each physical Lucene index (hereafter referred to as `partition`) will store document whose timestamp falls within the time range it covers. Partitions do *not* have overlapping time ranges.
A document's timestamp, for the purposes of partitioning, is indicated by a record's field name in the index's options.
The following options are added to the Lucene index:
- `textDocumentPartitionTimestamp`: the name of field to use as timestamp for partitioning. If this option isn't specified, the index is not partitioned.
- `textDocumentParititionHiWatermark`: high watermark size for a lucene partition, beyond which the partition would be split, or a new partition would be created
- `textDocumentParititionLoWatermark`: low watermark size for a lucene partition, below which the partition would be a candidate for merge with another adjacent partition.

The new partitions will be laid out in subspaces exactly like the current implementation, however they will be scoped as follows:
- `indexSubspace.subspace(Tuple.from(groupingKey, 1, partitionId))`
where `partitionId` is an integer starting at 0

Partition metadata will be stored in FDB as kv pairs at:
- `indexSubspace.subspace(Tuple.from(groupingKey, 0)`
the key is the `partitionId`, and the value is a protobuf message containing the `partitionId`, `from` timestamp, `to` timestamp, and document `count`

This update *does not* (yet) support
- splitting or merging partitions
- index merging across multiple partitions
- query continuation
- optimized query based on timestamp predicate (if provided)

This update sends all queries (including spellchecks) to the most recent partition