Fixes #2408: Add planner configuration property to control union ordering key behavior (#2409)
In order to avoid breaking continuations of plans affected by #2336, this adds a planner configuration flag that allows the user to effectively revert to the old behavior. This property is set to the old behavior by default, letting users to opt-in to the new behavior if they know it is safe for their deployment. This also includes a bit of a revamp of the or query tests to make sure all of the configurations get tested. There were two queries that had their plan hashes updated in #2350 (resolving #2336). The plan hashes in this PR for the legacy mode match the old plan hashes from before #2350 went in.

This fixes #2408.