Correctly use the resourceDescription for IndexInput
It's supposed to be more informative than just the filename, and
also be opaque. This changes `FDBIndexInput` to have a proper
parameter & field for the `fileName`, and pass the description on.
The key change that the test cared about is that the slice command
is supposed to include the parent input description and the slice,
which we can now do, because the `resourceDescription` isn't being
used as the file name.
This also means we can replace the logging parameter:
`nestedResourceDescription`, with a potentially more useful
`IndexInput requestingInput`, which, with the updated `slice` will
have all that information.