Remove exception swallowing
I ran this test, and, I believe it failed with a conflict, and
then proceeded to overwhelm the logging by logging ~15000 stack
traces that the transaction was no longer active.
And I think we should always be able to parse this file, or we
shouldn't silently run the test with 0 cities or something like
that.