Removed old coverage.gradle
I'm pretty sure this file is from a very old version of jacoco, and
substantially more complicated than it needs to be.
I tested locally with the property enabled, and it seemed to
gather jacoco data and generate reports.
Also, AFAICT we don't currently use this anywhere, so it would really
be up to a developer to run locally and assess the results.