Resolves #2431: Centralize the testing configuration, and other cleanup (#2423)
* Centralize testing configurations

This centralizes the testing config in testing.gradle, moving
the handling of SKIP_FDB_TESTS and SKIP_SLOW_TESTS there.
And:
- adds a default timeout to all tests of 2 seconds if you're
  skipping slow tests or 30 seconds if not
- adds a performanceTest task with the hopes of removing the
  process detecting code used before
- rearranges the tests for BlockingInAsyncDetection to make sure
  that the lucene configuration doesn't break the default configuration
- disables assertions and BlockingInAsyncDetection for performance
  tests
- removes jenkinsUnitTest; it wasn't used by anyone

* Try to fix jacoco failure

It failed :fdb-extensions:destructiveTest because it tried to
access testTask.jacoco but that didn't exist. Commenting out for
now, skipping the code in there, but probably will need to fix
if anyone is looking at jacoco results anyways.

* Disable blocking-in-async-detection for some tests

If the tests were reworked they could probably not disable the
detection, but that seemed like a bit of work that I don't want
to do right now.

* Remove exception swallowing

I ran this test, and, I believe it failed with a conflict, and
then proceeded to overwhelm the logging by logging ~15000 stack
traces that the transaction was no longer active.
And I think we should always be able to parse this file, or we
shouldn't silently run the test with 0 cities or something like
that.

* Fix dependency on unzipGeonames

I ran test and gradle complained that processTestResources had an
implicit dependency on unzipGeonames; since test depends on
processTestResources, updated this dependency to be explicit

* Removed old coverage.gradle

I'm pretty sure this file is from a very old version of jacoco, and
substantially more complicated than it needs to be.
I tested locally with the property enabled, and it seemed to
gather jacoco data and generate reports.
Also, AFAICT we don't currently use this anywhere, so it would really
be up to a developer to run locally and assess the results.

* Fix unzipGeonames dependency

It was writing to `.out` so anything that touched anything in
`.out` was spewing a warning. Changing it to a subdirectory solved
the issue.
(if the zip contained any directories overlapping with other tasks
that would have been an actual problem)

* Always run with jacoco & add jacoco-report-aggregation

* Create JacocoReport task for each test task, except performance

* Generate xml/html jacoco reports for all non-performance test tasks

* Remove jacoco-report-aggregation

This is apparently tricky to get working, instead let consumers
aggregate themselves.

* Exclude WipesFDB from regular test command

* Add missing RequiresFDB