Resolves #2455: sync before touching fieldInfos
FieldInfos breaks with NRTCachingDirectory

Found with some of the Lucene random tests, it won't be able to
find the reference without syncing the file first.
This might be unnecessary in production code, but sync is a no-op
in FDBDirectory.

This doesn't add the seed for the test mentioned in that issue,
because that now fails because of another issue, with a fix to come.

This also includes a similar fix for CompoundFormats needing to sync,
and a Seed for a test that that reproduces it
(LuceneOptimizedPointsFormatTest.testMultiValued with seed:
 [C185081D42F0F43C:FA63D2AE2DE7C6B0]).
I did not write a test for that, because writing an isolated test
for CompoundFormat seems much harder.