([SQL] Integrate basic permuted min / max indexes) (REL-628)
*  ([SQL] Integrate basic permuted min / max indexes)

This adds basic support for the permuted min and max index to Relational. It relies on changes to the Record Layer which are not yet in (https://github.com/FoundationDB/fdb-record-layer/pull/2420). With those changes, it allows us to use permuted min or max index in order to execute SQL queries with a `min` or `max` within it, including honoring order constraints.