Resolves #2479: Allow easily running lucene tests with randomness (#2448)
* Resolves #2479: Allow easily running lucene tests with randomness

This moves the seed for all the tests from an @Seed annotation
to a property in the build, and adds some gradle properties
to control whether to run with randomness. It also adds the
ability to run with multiple iterations, and run the nightly
tests.
In order to do this, I consolidated the test configuration, so
that destructiveTest, and test share more of the configuration.
Since the seed is provided in a different way for PRB, I changed
the way LuceneOptimizedDocValuesFormatTest.testMissingSortedBytes
avoids the problematic seed. This allows this test to still run
with a different seed from everyone else when not including random,
but also, if you include random it will run with a random seed.

* Add javadoc to the various test Tags

* allowAddIndexes in more tests that need it sometimes