Fixes #2485: FDBExceptions::wrapException incorrectly wraps InterruptedExceptions
This updates the logic in `FDBExceptions::wrapException` to avoid an error with the message "Unbalanced key/value logging info". This was ultimately caused by a problem where the wrong constructor was being selected when creating the `RecordCoreInterruptedException`, so this adds a cast so that the signature now aligns with the correct constructor.

This fixes #2485.