Resolves #2491: Use multiple partitions to perform rebalancing
The main change here is that this uses multiple transactions to
move documents between partitions. It also ensures that merge
happens on all partitions.
This is a rather simplistic implementation, with a fixed property
for how many to move per transaction.
In order to test all the permutations this adds some randomized tests
that create a bunch of random data, in a single, most-recent partition
and then runs the mergeIndex to move them between partitions.

It also does a few other things:
1. The search used in getOldestNDocuments was not right, and wasn't
   sorting properly
2. The JOINED index used in tests has the sort added
3. It changes the tuple at the beginnig by doing `groupKey.add()` rather
   than using a nested tuple.