Resolves #2472: Match repeated columns in grouped aggregate index (#2473)
* Resolves #2472: Match repeated columns in grouped aggregate index

The main change here is to improve the way that the aggregate index expansion visitor turns an aggregate index definition into a matchable defintion. This updates the logic from:

1. Expand the whole expression like a value index then
1. Re-expand the grouping columns (the part that was failing) to construct the `GroupBy` relational expression

Into:

1. Expand the whole expression like a value index then
1. Separate out those values into the grouping and grouped components

This avoids the logic that was throwing an exception before, and it correctly handles the usual pitfalls with nested repeated (i.e., ensuring that values either retain the same parent quantifier or get different ones depending on how the expression is defined). The rest of the code is then validating that we do the right thing, either by succesfully querying or correctly deciding _not_ to query, depending on whether we can execute it. Some of those new tests may end up getting updated if they underlying issue gets resolved.

This resolves #2472.