Resolves #2493: PlanOrderingKey too strict when computing order for In union
This takes basically the same approach that was taken for #2350, but this time applying it to `InUnion` plans rather than just regular union plans.

This change is guarded behind a planner configuration plan. This is necessary because the new algorithm can choose new ordering keys for the in-union, and so the user needs to be able to control its roll out. This is to mirror what we did with #2408.

This resolves #2493.