Fixes #2502: DeleteRecordsWhere should modify predicate for synthetic record types
When we delete all elements of a type, we need to modify the check so that it will remove all synthetic recorsd that reference the deleted base elements. This means modifying the predicate slightly. In particular, if there's a delete predicate like:

```java
Query.field("foo").equalsValue("bar")
```

And there's an unnested record type where the parent constituent is named `p`, then the correct predicate on the unnested record is:

```java
Query.field("p").matches(Query.field("foo").equalsValue("bar"))
```

This PR makes a few changes:

1. It makes sure that we gather up the indexes on synthetic types to check whether they are valid. This used to not work for cases where we were indexing only a single record type.
1. It validates that it can modify the predicate. For instance, if the index is on a single unnested record type, it can rewrite the predicate to be on the parent constituent. This check attempts to be a little conservative, completely skipping `JoinedRecordTypes`, which have more complicted semantics here

This fixes #2502.