Fixes #2505: Disabled indexes should not participate in deleteRecordsWhere
When checking and deleting indexes during `deleteRecordsWhere`, we do not need to look at disabled indexes, as they should not be maintained. This adds a filter for disabled indexes. In addition to being slightly more efficient, this also allows users an escape hatch if they add an index that does not support a `deleteRecordsWhere` operation that they want to do. In theory, the user can disable the index, perform the `deleteRecordsWhere`, and then rebuild the index.

This fixes #2505.