Fixes #2502: DeleteRecordsWhere should modify predicate for synthetic record types (#2508)
* refactor method to get matcher for delete where

* Fixes #2502: DeleteRecordsWhere should modify predicate for synthetic record types

When we delete all elements of a type, we need to modify the check so that it will remove all synthetic recorsd that reference the deleted base elements. This means modifying the predicate slightly. In particular, if there's a delete predicate like:

```java
Query.field("foo").equalsValue("bar")
```

And there's an unnested record type where the parent constituent is named `p`, then the correct predicate on the unnested record is:

```java
Query.field("p").matches(Query.field("foo").equalsValue("bar"))
```

This PR makes a few changes:

1. It makes sure that we gather up the indexes on synthetic types to check whether they are valid. This used to not work for cases where we were indexing only a single record type.
1. It validates that it can modify the predicate. For instance, if the index is on a single unnested record type, it can rewrite the predicate to be on the parent constituent. This check attempts to be a little conservative, completely skipping `JoinedRecordTypes`, which have more complicted semantics here

This fixes #2502.

* Fixes #2505: Disabled indexes should not participate in deleteRecordsWhere

When checking and deleting indexes during `deleteRecordsWhere`, we do not need to look at disabled indexes, as they should not be maintained. This adds a filter for disabled indexes. In addition to being slightly more efficient, this also allows users an escape hatch if they add an index that does not support a `deleteRecordsWhere` operation that they want to do. In theory, the user can disable the index, perform the `deleteRecordsWhere`, and then rebuild the index.

This fixes #2505.

* Add a new test for PERMUTED_MIN and PERMUTED_MAX indexes in FDBRecordStoreDeleteWhereTest

There's actually already a test for `deleteRecordsWhere` within `PermutedMinMaxIndexTest`, though I added a test case that checks to see if we are stopped if we try to delete what appears to be a prefix of groups in the index, but we actually can't because of how the `permutedSize` option messes with things.

* Add coverage for and components in delete where with unnested constituents

* remove extraneous import

* remove now extraneous check from deleteRecordsWhere