Fixex #2513: Extra errors being reported by the scrubber for indexes on Unnested Record Types
Adds some additional tests for index building of the unnested and joined record types. This includes updates to the OnlineIndexerBuildIndexTest so that it can be used for more index types than just MySimpleRecord.

Also makes changes to the scrubbers so that they have explicit handling of synthetic record types. There is probably some additional work that will be needed to make this sure that we've covered all of the interesting corner cases.

Another thing that came up during testing: during check version, we were not quite validating the right thing. We had been trying to scan the synthetic record's "primary key range" by looking for all records beginning with the synthetic record's type key. That meant that we'd often mark the index as already readable. This makes changes to how the types are constructed for the `getIndexesSince` method so that it only returns stored types. I added some tests to the `UnnestedRecordTypeTest` to validate that the new index gets marked as readable if and only if the index is on new (stored) record types.