Fixes #2530: Synthetic records use wrong configuration if built during checkVersion
This updates the logic used to build indexes during check version so that it no longer takes the record type list from `getIndexesSince`. This means that the online indexer recalculates the list of record types it needs, which allows it to properly note whether the index is on synthetic records or not. The new test used to fail before this change.

This fixes #2350.