Fixes #2543: Permuted min/max indexes on keys producing duplicates only index the first entry
The core issue was that there was an early-out in the `PermutedMinMaxIndexMaintainer`'s implementation of `updateIndexKeys`. This updates the loop so that it no longer early outs, and thus all entries will get included in the index. Some care has been taken so that if the same record produces multiple entries in the same group (which can happen if the repeated field is in the grouping key and the same value appears multiple time in the repeated field OR if the repeated field is in the aggregate value), then only the most extreme value for the record is used to determine what new entries to add to the permuted space. All values still get added to the primary index space, though, so that part of the index should behave exactly like a normal value index on the whole key (i.e., ignoring the `GroupingKeyExpression`).

This fixes #2543.