Fixes #2552: Nested fields can preserve field ordinals incorrectly in index expansion
This updates the way that graph expansions get combined during index match candidate generation so that field ordinals are not preserved. There's in general somthing iffy about the way that field ordinals are being treated here, and it's possible there are some other locations where we're moving the `Field` object around incorrectly.

This also introduces a new helper method to make it easier to write `Column`s without needing to repeat the type and uses that new abstraction in some existing code paths.

This fixes #2552.