(Final Integration for SQL Bundle Queries) (REL-677)
This does two things. The first is it combines the two bundle query yamsqls into one. This allows the schema with the repeated fields that were added for `bundle-query-1.yamsql` to be re-used by both bundle query tests. The second thing it does is add additional tests for the bundle query 1 which make use of the permuted max index, in particular adding tests that combine max predicates with the map field so that we have tests for exactly the kind of query that will be sent for these. That uncovered a number of ancillary issues that this also attempts to address.

 (Final Integration for SQL Bundle Queries)