Fixes #2588: Index planning can match against nested synthetic type primary key components incorrectly
This adjusts the way the planner interacts with primary keys using a `ListKeyExpression` to avoid matching them incorrectly during an index scan. There was an additional problem, related to matching the `RecordTypeKeyExpression`, where the physical operator had a bug in that it didn't resolve the record type in a way that handled synthetic record types, and that bug has also been addressed.

This fixes #2587 and this fixes #2588.