Fixes #2600: LuceneIndexMaintenanceTest.flakyMerge passes
Technically this just updates the test, the previous
commits change the code, and together the test passes.
The test only runs some of the configurations during PRB,
because each configuration takes a couple minutes.
Other configurations are written if `tests.nightly` is enabled.
Doing so required cleaning up the lucene test framework
so that its nightly tests that fail are disabled.