FDBDirectory clear file metadata after file data
In the case of an Agile Context, this will ensure that the file
reference sticks around if that is in a different context from the
rest of the deletes, and only some of the deletes succeed.
Having the reference around means that lucene will come back and
see an unreferenced file and clean it up, causing us to retry
cleaning all the other data.