Resolevs #2601: Tests should run in unique key space paths
This introduces a new class, `TestKeySpacePathManager`, which does two things:

1. It allocates new key space paths for tests. Each newly allocated key space path contains a UUID that should make it unique for each test.
1. When the manager gets closed, it deletes the data from any key spaces that it has allocated.

This then provides a `TestKeySpacePathManagerExtension` which wraps the new class and makes sure that the path manager gets closed in an `AfterEach` callback. This also required reworking the way that the `FDBDatabase` gets initialized from the factory to use a testing extension, which in turn makes it harder to do something like incidentally use an `FDBDatabase` with the wrong API version set.

This path manager class is meant to be generic enough that if we needed to, we could theoretically allocate one keyspace for a single test _class_ and then use this same abstraction to handle its lifecycle, though that is not something that this attempts to handle. That's why the only extension that was added has a lifecycle that conforms with each test _method_ getting its own key space path. However, the path manager abstraction is supposed to be generic enough that we could theoretically extend it so that it manages a key space path which is shared among all tests in a given class.

This also does something similar with a `TestDatabaseExtension` and a `TestSubspaceExtension` for use within the `fdb-extensions` tests. Those could theoretically be spruced up a little, but they operate under similar principles.

This resolves #2601.